brincando-com-imagens-e-streamlit/requirements.txt
altair==4.1.0; python_version >= "3.6"
appnope==0.1.2; platform_system == "Darwin" and python_version >= "3.7" and sys_platform == "darwin"
argon2-cffi==21.1.0; python_version >= "3.7"
astor==0.8.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
attrs==21.2.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
backcall==0.2.0; python_version >= "3.7"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.6"
base58==2.1.0; python_version >= "3.6"
bleach==4.1.0; python_version >= "3.7"
blinker==1.4; python_version >= "3.6"
cachetools==4.2.4; python_version >= "3.6" and python_version < "4.0"
certifi==2021.5.30; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
cffi==1.14.6; python_full_version >= "3.6.1" and python_version >= "3.7" and implementation_name == "pypy"
charset-normalizer==2.0.6; python_full_version >= "3.6.0" and python_version >= "3.6"
click==7.1.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0"
debugpy==1.5.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
decorator==5.1.0; python_version >= "3.7"
defusedxml==0.7.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
entrypoints==0.3; python_full_version >= "3.6.1" and python_version >= "3.7"
gitdb==4.0.7; python_version >= "3.7"
gitpython==3.1.24; python_version >= "3.7"
idna==3.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
ipykernel==6.4.1; python_version >= "3.7"
ipython-genutils==0.2.0; python_version >= "3.7"
ipython==7.28.0; python_version >= "3.7"
ipywidgets==7.6.5; python_version >= "3.7"
jedi==0.18.0; python_version >= "3.7"
jinja2==3.0.1; python_version >= "3.7"
jsonschema==4.0.1; python_version >= "3.7"
jupyter-client==7.0.5; python_full_version >= "3.6.1" and python_version >= "3.7"
jupyter-core==4.8.1; python_full_version >= "3.6.1" and python_version >= "3.7"
jupyterlab-pygments==0.1.2; python_version >= "3.7"
jupyterlab-widgets==1.0.2; python_version >= "3.7"
markupsafe==2.0.1; python_version >= "3.6"
matplotlib-inline==0.1.3; python_version >= "3.7"
mistune==0.8.4; python_version >= "3.7"
nbclient==0.5.4; python_full_version >= "3.6.1" and python_version >= "3.7"
nbconvert==6.2.0; python_version >= "3.7"
nbformat==5.1.3; python_full_version >= "3.6.1" and python_version >= "3.7"
nest-asyncio==1.5.1; python_full_version >= "3.6.1" and python_version >= "3.7"
notebook==6.4.4; python_version >= "3.7"
numpy==1.21.1; python_full_version >= "3.7.1" and python_version >= "3.7"
packaging==21.0; python_version >= "3.7"
pandas==1.3.3; python_full_version >= "3.7.1" and python_version >= "3.6"
pandocfilters==1.5.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
parso==0.8.2; python_version >= "3.7"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.7"
pickleshare==0.7.5; python_version >= "3.7"
pillow==8.3.2; python_version >= "3.6"
prometheus-client==0.11.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
prompt-toolkit==3.0.20; python_full_version >= "3.6.2" and python_version >= "3.7"
protobuf==3.18.0; python_version >= "3.6"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.7" and os_name != "nt"
py==1.10.0; python_full_version >= "3.6.1" and python_version >= "3.7" and implementation_name == "pypy"
pyarrow==5.0.0; python_version >= "3.6"
pycparser==2.20; python_full_version >= "3.6.1" and python_version >= "3.7" and implementation_name == "pypy"
pydeck==0.7.0; python_version >= "3.7"
pygments==2.10.0; python_version >= "3.7"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
pyrsistent==0.18.0; python_version >= "3.7"
python-dateutil==2.8.2; python_full_version >= "3.7.1" and python_version >= "3.7"
pytz==2021.3; python_full_version >= "3.7.1" and python_version >= "3.6"
pywin32==301; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_full_version >= "3.6.1" and python_version >= "3.7"
pywinpty==1.1.4; os_name == "nt" and python_version >= "3.7"
pyzmq==22.3.0; python_full_version >= "3.6.1" and python_version >= "3.7"
requests==2.26.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
send2trash==1.8.0; python_version >= "3.7"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
smmap==4.0.0; python_version >= "3.7"
streamlit==0.89.0; python_version >= "3.6"
terminado==0.12.1; python_version >= "3.7"
testpath==0.5.0; python_version >= "3.7"
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
toolz==0.11.1; python_version >= "3.6"
tornado==6.1; python_full_version >= "3.6.1" and python_version >= "3.7"
traitlets==5.1.0; python_full_version >= "3.6.1" and python_version >= "3.7"
typing-extensions==3.10.0.2; python_version < "3.10" and python_version >= "3.7"
tzdata==2021.2.post0; platform_system == "Windows" and python_version >= "3.6"
tzlocal==3.0; python_version >= "3.6"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
validators==0.18.2; python_version >= "3.6"
watchdog==2.1.6; platform_system != "Darwin" and python_version >= "3.6"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.7"
webencodings==0.5.1; python_version >= "3.7"
widgetsnbextension==3.5.1; python_version >= "3.7"
